fixed remove item render and removed cart files
