+++ client/redux/action-creators/order.js
@@ -1,4 +1,10 @@
-import {GET_USER_ORDER, ADD_PRODUCT} from '../action-types/order'
+import {
+  INCREMENT_QTY,
+  DECREMENT_QTY,
+  REMOVE_ITEM,
+  GET_USER_ORDER,
+  ADD_PRODUCT
+} from '../action-types/order'
 
 export const getUserOrder = userOrder => {
   return {
@@ -10,6 +16,27 @@ export const getUserOrder = userOrder => {
 export const addProduct = product => {
   return {
     type: ADD_PRODUCT,
+    product
+  }
+}
+
+export const incrementQty = product => {
+  return {
+    type: INCREMENT_QTY,
+    product
+  }
+}
+
+export const decrementQty = product => {
+  return {
+    type: DECREMENT_QTY,
+    product
+  }
+}
+
+export const removeItem = product => {
+  return {
+    type: REMOVE_ITEM,
     product
   }
 }
