--- client/redux/action-types/cart.js
@@ -1,3 +1,0 @@
-export const DECREMENT_QTY = 'DECREMENT_QTY'
-export const INCREMENT_QTY = 'INCREMENT_QTY'
-export const REMOVE_ITEM = 'REMOVE_ITEM'
