+++ client/redux/reducers/order.js
@@ -1,5 +1,10 @@
-import {GET_USER_ORDER, ADD_PRODUCT} from '../action-types/order'
-// test
+import {
+  GET_USER_ORDER,
+  ADD_PRODUCT,
+  REMOVE_ITEM,
+  DECREMENT_QTY,
+  INCREMENT_QTY
+} from '../action-types/order'
 const initialState = {
   userOrder: {}
 }
@@ -9,10 +14,56 @@ export const order = (state = initialState, action) => {
     case GET_USER_ORDER:
       return {...state, userOrder: action.userOrder}
     case ADD_PRODUCT:
-      // console.log('state', state, 'state products', state.userOrder.products)
+      console.log('state products', state.userOrder.products)
       return {
         ...state,
-        [state.products]: [...state.userOrder.products, action.product]
+        userOrder: {
+          ...state.userOrder,
+          products: [...state.userOrder.products, action.product]
+        }
+      }
+    case REMOVE_ITEM:
+      return {
+        ...state,
+        userOrder: {
+          ...state.userOrder,
+          products: state.userOrder.products.filter(
+            product => product.id !== action.product.id
+          )
+        }
+      }
+
+    case DECREMENT_QTY:
+      return {
+        ...state,
+        products: state.userOrder.products.map(product => {
+          if (product.id === action.product.id) {
+            return {
+              ...product,
+              orderItem: {
+                ...product.orderItem,
+                quantity: product.orderItem.quantity - 1
+              }
+            }
+          }
+          return product
+        })
+      }
+    case INCREMENT_QTY:
+      return {
+        ...state,
+        products: state.userOrder.products.map(product => {
+          if (product.id === action.product.id) {
+            return {
+              ...product,
+              orderItem: {
+                ...product.orderItem,
+                quantity: product.orderItem.quantity + 1
+              }
+            }
+          }
+          return product
+        })
       }
     default:
       return state
