+++ client/redux/store/index.js
@@ -3,11 +3,10 @@ import {createLogger} from 'redux-logger'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import {user} from '../reducers/user'
 import {products} from '../reducers/products'
-import {cart} from '../reducers/cart'
 import {order} from '../reducers/order'
 import thunkMiddleware from 'redux-thunk'
 
-const reducer = combineReducers({user, products, cart, order})
+const reducer = combineReducers({user, products, order})
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
